XXX PATCH: [perl #120723] Setting LC_NUMERIC breaks parsing of constants
LC_NUMERIC is handled differently than the other locale categories,
in that for this category, Perl normally stays in the C locale,
converting to the underlying locale for brief periods, only when
necessary.  This is undocumented (until the commit following this one).

The problem was that when POSIX::setlocale() is called, it changes the
locale, which is correct only if called during those places where the
underlying locale is supposed to be in effect, or if the new locale is
the C locale.

This commit simply adds code to change the locale back to the C locale
if appropriate.